+++ CHANGELOG.md
@@ -7,8 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
 - **CUMULUS-2717**
   - Added `Show/Hide Recovery Status` button to granule detail page
+- **CUMULUS-2915**
+  - Added download button to execution pages
+### Changed
+
 - **CUMULUS-2763**
   - Replaces disabled button interaction with hidden buttons for bulk actions.
 - **CUMULUS-2960**